favorites-collections-service/requirements.txt
# Favorites & Collections Service Dependencies
# Python 3.10+ required

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Async PostgreSQL support
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.25

# Validation and settings
pydantic>=2.5.0
pydantic-settings>=2.1.0

# JWT authentication
PyJWT>=2.8.0

# Date/time handling (for timezone-aware datetimes)
python-dateutil>=2.8.2

# Development dependencies (optional)
# Uncomment for development
# pytest>=8.0.0
# pytest-asyncio>=0.23.0
# httpx>=0.26.0  # For async test client
# black>=24.1.0
# isort>=5.13.0
# mypy>=1.8.0


